Move user resources to own package
