+++ pkg/artifactory/resource/user/resource_artifactory_anonymous_user.go
@@ -1,4 +1,4 @@
-package artifactory
+package user
 
 import (
 	"context"
