+++ pkg/artifactory/resource/user/resource_artifactory_anonymous_user_test.go
@@ -1,4 +1,4 @@
-package artifactory_test
+package user_test
 
 import (
 	"fmt"
