+++ pkg/artifactory/resource/user/resource_artifactory_user.go
@@ -1,4 +1,4 @@
-package artifactory
+package user
 
 import (
 	"context"
