+++ pkg/artifactory/resource/user/user.go
@@ -1,4 +1,4 @@
-package artifactory
+package user
 
 import (
 	"context"
